+++ changelog.md
@@ -1,12 +1,18 @@
 # Changelog
 
+## [1.0.4] - 2024-06-12
+
+### Added
+- `parse` to convert a createHash string to an array of bytes.
+- `stringify` to convert an array of bytes to a createHash string.
+- `validate` to check if a string is a valid createHash.
+
 ## [1.0.3] - 2024-06-06
 
 ### Added
 - `createHash()`: Generates a unique key with "Arlecchino" characters.
 - `createManyHashes(count)`: Generates the specified number of unique keys.
 - `createCustomHash(customString)`: Generates a custom key with the specified string.
-- Enhanced unit tests to cover new functionalities.
 
 ## [1.0.1] - 2024-05-28